pyproject.toml
[tool.poetry]
name = "navigator-data-ingest"
version = "0.1.0"
description = "Loads data into backend"
authors = ["CPR tech <tech@climatepolicyradar.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.2"
pydantic = "^1.9"
python-slugify = "^6.1.2"
python-dateutil = "^2.8.2"
requests = "^2.28.1"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
importlib-resources = "^5.6.0"
pre-commit = "^2.17.0"
pydocstyle = { version = "^6.1.1", extras = ["pydocstyle[toml]"] }
pytest = "^7.0.1"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
asyncio_mode = "strict"

[tool.pydocstyle]
# D100 - Missing docstring in public module
# D103 - Missing docstring in public function
# D104 - Missing docstring in public package
# D107 - Missing docstring in __init__
# D202 - No blank lines allowed after function docstring
# D203 - 1 blank line required before class docstring
# D213 - Multi-line docstring summary should start at the second line
# D400 - First line should end with a period
# D401 - First line should be in imperative mood (Google styleguide denotes 'descriptive' style.)
# D406 - Section name should end with a newline (incompatible with Google styleguide)
# D407 - Missing dashed underline after section (incompatible with Google styleguide)
# D413 - Missing blank line after last section (incompatible with Google styleguide)
# D415 - First line should end with a period, question mark, or exclamation point
ignore = """
D100,D103,D104,D107,
D202,D203,D212,
D400,D401,D406,D407,D413,D415
"""

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
per-file-ignores = "__init__.py:F401"

[tool.black]
line-length = 88
target-version = ['py39']
